Multiplayer game called Die.
This project is based on a project by Todd Neller of Gettysburg College. In this project you would implement a multiplayer game called Die.

Die is a multiplayer dice game with simple rules: Each player races to reach a goal of 100 points. On each turn, a player repeatedly rolls a die until either a 1 ("die") is rolled or the player holds and scores the sum of the rolls (which becomes the turn total).

At any time during a player's turn, the player is faced with two decisions:
roll - If the player rolls a
1: the player scores nothing and it becomes the opponent's turn.
2 - 6: the number is added to the player's turn total and the player's turn continues.
hold - The turn total is added to the player's score and it becomes the opponent's turn.